+++ CHANGELOG.md
@@ -13,6 +13,18 @@
 ### Fixed
 
 ### Security
+## [unspecified]
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
 
 ## [0.0.5]
 ### fixed