+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 ### Added
 - ansible json schema's for validation
+- Added json schema's for Ansible
 
 ### Fixed
 - changelog updates to published plugin so it can be seen within idea marketplace